GUI done, should be checked

+++ Client/RequestsHandler.cs
@@ -6,7 +6,6 @@ using SRTShareLib.SRTManager.Encryption;
 using SRTShareLib.SRTManager.RequestsFactory;
 using System;
 using System.Diagnostics;
-using System.Runtime.CompilerServices;
 using System.Windows.Forms;
 
 using CConsole = SRTShareLib.CColorManager;
@@ -35,7 +34,7 @@ namespace Client
             else
                 myPublicKey = new byte[DiffieHellman.PUBLIC_KEY_SIZE];
 
-            Packet handshake_packet = handshake_response.Conclusion(init_psn: LiveStream.INITIAL_PSN, p_ip: peer_ip, clientSide: true, LiveStream.My_SID, handshake_request.SOURCE_SOCKET_ID, handshake_request.ENCRYPTION_TYPE, myPublicKey, handshake_request.RETRANSMISSION_MODE);
+            Packet handshake_packet = handshake_response.Conclusion(init_psn: (uint)LiveStream.INITIAL_PSN, p_ip: peer_ip, clientSide: true, LiveStream.My_SID, handshake_request.SOURCE_SOCKET_ID, handshake_request.ENCRYPTION_TYPE, myPublicKey, handshake_request.RETRANSMISSION_MODE);
             PacketManager.SendPacket(handshake_packet);
         }
 
@@ -53,7 +52,7 @@ namespace Client
 
             CConsole.WriteLine("[Handshake completed] Starting video & audio transmission\n", MessageType.bgSuccess);
             AudioPlay.PrepareAudio();  // init objects before receiving audio
-            
+
 
             EnableQualityButtons(mainView);
         }
@@ -70,7 +69,7 @@ namespace Client
                     (OSIManager.BuildBaseLayers(NetworkManager.MacAddress, server_mac, NetworkManager.LocalIp, ConfigManager.IP, myPort, ConfigManager.PORT));
 
             IpV4Address peer_ip = new IpV4Address(LiveStream.GetAdaptedIP());
-            Packet handshake_packet = handshake.Induction(init_psn: LiveStream.INITIAL_PSN, p_ip: peer_ip, clientSide: true, client_socket_id, 0, (ushort)LiveStream.ENCRYPTION, new byte[DiffieHellman.PUBLIC_KEY_SIZE], LiveStream.RETRANSMISSION_MODE);
+            Packet handshake_packet = handshake.Induction(init_psn: (uint)LiveStream.INITIAL_PSN, p_ip: peer_ip, clientSide: true, client_socket_id, 0, (ushort)LiveStream.ENCRYPTION, new byte[DiffieHellman.PUBLIC_KEY_SIZE], LiveStream.RETRANSMISSION_MODE);
 
             PacketManager.SendPacket(handshake_packet);
         }
